+++ RELEASES.md
@@ -551,3 +551,6 @@ based on happner 1.28.1
 10.5.0 2019-10-03
 -----------------
   - security component listUsers and listGroups with criteria
+  - security component listUsers and listGroups with count, skip and limit
+  - mongo plugin aggregate, collation and count
+  - fix: uglify-es in packager